requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi~=2022.12.7
greenlet==3.0.1
idna==3.4
numpy~=1.24.2
pandas~=2.0.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz~=2023.3
requests~=2.28.2
schedule==1.2.1
six~=1.15.0
soupsieve==2.5
SQLAlchemy==2.0.23
urllib3~=1.26.15
lxml~=4.9.2
pip~=23.3.1
attrs~=23.1.0
wheel~=0.41.3
filelock~=3.13.1
setuptools~=68.2.2
MarkupSafe~=1.1.1
simplejson~=3.19.2
future~=0.18.2
platformdirs~=3.11.0
python-dotenv~=1.0.0
